requirements.txt
requests==2.27.1
python-dotenv==0.20.0
pylint==2.15.10